finished refactoring of packages
