+++ src/test/java/at/uibk/leco/service/RoomTableServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package at.uibk.leco.service;
 
 import at.uibk.leco.models.Room;
 import at.uibk.leco.models.RoomTable;
