+++ src/test/java/at/uibk/leco/service/TimingServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package at.uibk.leco.service;
 
 import at.uibk.leco.constants.TimingConstants;
 import at.uibk.leco.models.RoomTable;
